pyproject.toml
[project]
name = "openai-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "function-tool>=0.0.4",
    "openai-agents>=0.2.8",
    "pymongo[srv]>=4.14.1",
    "python-dotenv>=1.1.1",
]
